# pkg/api/openapi/notpastebin.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "notpastebin.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "NotPasteBin"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/note/create": {
      "post": {
        "operationId": "NotPasteBin_CreateNote",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiCreateNoteResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/apiCreateNoteRequest"
            }
          }
        ],
        "tags": [
          "NotPasteBin"
        ]
      }
    },
    "/note/delete": {
      "post": {
        "operationId": "NotPasteBin_DeleteNote",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "type": "object",
              "properties": {}
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/apiDeleteNoteRequest"
            }
          }
        ],
        "tags": [
          "NotPasteBin"
        ]
      }
    },
    "/note/get": {
      "post": {
        "operationId": "NotPasteBin_GetNote",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiGetNoteResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/apiGetNoteRequest"
            }
          }
        ],
        "tags": [
          "NotPasteBin"
        ]
      }
    },
    "/note/update": {
      "post": {
        "operationId": "NotPasteBin_UpdateNote",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "type": "object",
              "properties": {}
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/apiUpdateNoteRequest"
            }
          }
        ],
        "tags": [
          "NotPasteBin"
        ]
      }
    }
  },
  "definitions": {
    "apiCreateNoteRequest": {
      "type": "object",
      "properties": {
        "text": {
          "type": "string"
        }
      }
    },
    "apiCreateNoteResponse": {
      "type": "object",
      "properties": {
        "adminHash": {
          "type": "string"
        },
        "readerHash": {
          "type": "string"
        }
      }
    },
    "apiDeleteNoteRequest": {
      "type": "object",
      "properties": {
        "adminHash": {
          "type": "string"
        }
      }
    },
    "apiGetNoteRequest": {
      "type": "object",
      "properties": {
        "hash": {
          "type": "string"
        }
      }
    },
    "apiGetNoteResponse": {
      "type": "object",
      "properties": {
        "text": {
          "type": "string"
        },
        "isAdmin": {
          "type": "boolean"
        },
        "readerHash": {
          "type": "string"
        }
      }
    },
    "apiUpdateNoteRequest": {
      "type": "object",
      "properties": {
        "text": {
          "type": "string"
        },
        "adminHash": {
          "type": "string"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}